[sensorDB] remove 'exif' module
- new module for the sensor width database 'sensorDB'
