--- src/aliceVision/exif/ExifIO.cpp
@@ -1,55 +1,0 @@
-// This file is part of the AliceVision project.
-// This Source Code Form is subject to the terms of the Mozilla Public License,
-// v. 2.0. If a copy of the MPL was not distributed with this file,
-// You can obtain one at https://mozilla.org/MPL/2.0/.
-
-#include "ExifIO.hpp"
-
-#include "aliceVision/stl/hash.hpp"
-
-#include <cstdlib>
-
-namespace aliceVision {
-namespace exif  {
-
-std::size_t computeUID(const ExifIO& exifReader, const std::string& imageFilename)
-{
-  std::size_t uid = 0;
-
-  if( !exifReader.getImageUniqueID().empty() ||
-      !exifReader.getSerialNumber().empty() ||
-      !exifReader.getLensSerialNumber().empty()
-    )
-  {
-    stl::hash_combine(uid, exifReader.getImageUniqueID());
-    stl::hash_combine(uid, exifReader.getSerialNumber());
-    stl::hash_combine(uid, exifReader.getLensSerialNumber());
-  }
-  else
-  {
-    // No metadata to identify the image, fallback to the filename
-    stl::hash_combine(uid, imageFilename);
-  }
-  
-  if( !exifReader.getDateTimeOriginal().empty() )
-  {
-    stl::hash_combine(uid, exifReader.getDateTimeOriginal());
-    stl::hash_combine(uid, exifReader.getSubSecTimeOriginal());
-  }
-  else
-  {
-    // If no original date/time, fallback to the file date/time
-    stl::hash_combine(uid, exifReader.getDateTime());
-  }
-
-  stl::hash_combine(uid, exifReader.getWidth());
-  stl::hash_combine(uid, exifReader.getHeight());
-  
-  // Limit to integer to maximize compatibility (like Alembic in Maya)
-  uid = std::abs((int) uid);
-
-  return uid;
-}
-
-} // namespace exif
-} // namespace aliceVision
