+++ src/aliceVision/keyframe/CMakeLists.txt
@@ -19,6 +19,7 @@ target_include_directories(aliceVision_keyframe
 
 target_link_libraries(aliceVision_keyframe
   PUBLIC aliceVision_feature
+  aliceVision_sensorDB
   aliceVision_dataio
   aliceVision_voctree
   ${OPENIMAGEIO_LIBRARIES}
