+++ src/aliceVision/sensorDB/sensor_width_camera_database.txt
