--- src/dependencies/easyexif/CONTRIBUTORS
@@ -1,10 +1,0 @@
-The following people have contributed patches or issues to EasyEXIF:
-
-Seth Fowler
-Val Malykh
-Carlos Apablaza Brito
-Simon Fuhrmann
-Toshiaki Ohkuma
-pet.b.hunt
-Jason Moey
-Tomas Volf
