--- src/dependencies/easyexif/README.md
@@ -1,5 +1,0 @@
-Project: easyexif
-URL: https://github.com/mayanklahiri/easyexif
-License: New BSD License
-Upstream version: version 1.0 7e60481
-
